fix header-stock
