+++ client/Pages/MainBoard.jsx
@@ -32,6 +32,8 @@ const styles = {
 };
 
 function MainBoard() {
+  console.log('---');
+
   return (
     <div style={styles.wrapper}>
       <Switch>
