+++ client/Pages/Stock/StockPage.jsx
@@ -2,10 +2,15 @@
 /** @jsx jsx */
 
 import {
+  useEffect,
   useState,
   useMemo,
 } from 'react';
+import {
+  useParams,
+} from 'react-router-dom';
 import { jsx, css } from '@emotion/core';
+import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import {
   LineChart, Line,
@@ -14,6 +19,8 @@ import StockStrategyHeader from './StockStrategyHeader';
 import {
   incomeStatements, balanceSheets, cashFlows, dividends,
 } from '../../Constant/stockTable';
+import prettifyStockData from '../../helper/stocks';
+import * as StockActions from '../../actions/Stocks';
 
 const styles = {
   wrapper: css`
@@ -166,15 +173,43 @@ const TABLE_TYPES = {
 
 type Props = {
   stockData: Object,
-  state: Object,
+  storeStockData: Function,
 };
 
 function StockPage({
   stockData,
+  storeStockData,
 }: Props) {
+  const [table, setTable] = useState('INCOME_STATEMENT');
+
   console.log('stockData', stockData);
 
-  const [table, setTable] = useState('INCOME_STATEMENT');
+  const { stockId } = useParams();
+
+  useEffect(() => {
+    let canceled = false;
+
+    async function fetchStockData() {
+      const resData = await fetch(`${API_HOST}/stocker/individualStock/${stockId}`, {
+        method: 'GET',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      }).then(res => (!canceled ? res.json() : null));
+
+      if (resData) {
+        const prettifiedStockData = prettifyStockData(resData);
+
+        storeStockData(prettifiedStockData);
+      }
+    }
+
+    fetchStockData();
+
+    return () => {
+      canceled = true;
+    };
+  }, [storeStockData, stockId]);
 
   const infoTable = useMemo(() => {
     switch (table) {
@@ -429,6 +464,9 @@ const reduxHook = connect(
   state => ({
     stockData: state.Stocks.stockData,
   }),
+  dispatch => bindActionCreators({
+    ...StockActions,
+  }, dispatch),
 );
 
 export default reduxHook(StockPage);
